Feat: community 개발 | S03P22A111-61
- 수정, 삭제 권한 추가

+++ backend/community/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import get_object_or_404
 from rest_framework.response import Response
+from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.decorators import api_view
 from .models import Board, Favorite, Article, Comment
@@ -73,18 +74,24 @@ class ArticleDetail(APIView):
   # 게시물 update - params : title, content
   def put(self, request, board_pk, article_pk):
     article = Article.objects.get(pk=article_pk)
-    board = Board.objects.get(pk=board_pk)
-    serializer = ArticleCreateSerializer(article, data=request.data)
-    if serializer.is_valid(raise_exception=True):
-        serializer.save(user=request.user, board=board)
-        return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    if request.user == article.user: 
+      board = Board.objects.get(pk=board_pk)
+      serializer = ArticleCreateSerializer(article, data=request.data)
+      if serializer.is_valid(raise_exception=True):
+          serializer.save(user=request.user, board=board)
+          return Response(serializer.data)
+      return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    else:
+      return Response(status=status.HTTP_403_FORBIDDEN)
 
   # 게시물 delete
   def delete(self, request, board_pk, article_pk):
       article = Article.objects.get(pk=article_pk)
-      article.delete()
-      return Response(status=200)
+      if request.user == article.user:
+        article.delete()
+        return Response(status=200)
+      else:
+        return Response(status=status.HTTP_403_FORBIDDEN)
 
   # 댓글 create - params : content
   def post(self, request, board_pk, article_pk):
@@ -106,18 +113,24 @@ class CommentDetail(APIView):
   # 댓글 update
   def put(self, request, board_pk, article_pk, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    serializer = CommentCreateSerializer(comment, data=request.data)
-    article = get_object_or_404(Article, pk=article_pk)
-    if serializer.is_valid(raise_exception=True):
-      serializer.save(user=request.user, article=article)
-      return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    if request.user == comment.user:
+      serializer = CommentCreateSerializer(comment, data=request.data)
+      article = get_object_or_404(Article, pk=article_pk)
+      if serializer.is_valid(raise_exception=True):
+        serializer.save(user=request.user, article=article)
+        return Response(serializer.data)
+      return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    else:
+      return Response(status=status.HTTP_403_FORBIDDEN)
 
   # 댓글 delete
   def delete(self, request, board_pk, article_pk, comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return Response(status=200)
+    if request.user == comment.user:
+      comment.delete()
+      return Response(status=200)
+    else:
+      return Response(status=status.HTTP_403_FORBIDDEN)
 
 
 @api_view(['GET'])
